[None][test] Fold multimodal dummy-sizing tests into model test files
Move the Qwen-VL and Mistral3 deterministic dummy-sizing unit tests out of the
standalone tests/unittest/inputs/ files and into the existing model test files
(test_modeling_qwen2_5vl.py, test_modeling_mistral.py), co-locating them with
the models they cover. The tests stay CPU-only (stubbed input processors via
__new__ + SimpleNamespace, no model load), so no GPU dependency is added.
Removes the now-redundant standalone test files.
